Add logic to move bodies apart if they are intersecting.

+++ common.h
@@ -1,4 +1,4 @@
 #pragma once
 
-constexpr uint32_t SCREEN_WIDTH = 1200;
-constexpr uint32_t SCREEN_HEIGHT = 800;
+constexpr size_t SCREEN_WIDTH = 1200;
+constexpr size_t SCREEN_HEIGHT = 800;
